# backend/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Liveprofit API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios",
        "operationId": "list_portfolios_api_v1_portfolios_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PortfolioListData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "operationId": "create_portfolio_api_v1_portfolios_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PortfolioDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}": {
      "patch": {
        "tags": [
          "portfolios"
        ],
        "summary": "Update Portfolio",
        "operationId": "update_portfolio_api_v1_portfolios__portfolio_id__patch",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PortfolioDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Delete Portfolio",
        "operationId": "delete_portfolio_api_v1_portfolios__portfolio_id__delete",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "expected_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DeleteResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/positions": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Positions",
        "operationId": "list_positions_api_v1_portfolios__portfolio_id__positions_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PortfolioPositionsData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/positions/{market}/{symbol}": {
      "put": {
        "tags": [
          "portfolios"
        ],
        "summary": "Upsert Position",
        "operationId": "upsert_position_api_v1_portfolios__portfolio_id__positions__market___symbol__put",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "market",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Market"
            }
          },
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PortfolioPositionMutationData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Remove Position",
        "operationId": "remove_position_api_v1_portfolios__portfolio_id__positions__market___symbol__delete",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "market",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Market"
            }
          },
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "expected_portfolio_revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Portfolio Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DeleteResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteResultData": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "default": true
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id",
            "description": "被删除资源 UUID"
          }
        },
        "type": "object",
        "required": [
          "resource_id"
        ],
        "title": "DeleteResultData"
      },
      "EnvelopeMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "请求标识（X-Trace-ID）"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "v1"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "不透明续页游标；无下一页为 null"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "EnvelopeMeta"
      },
...
      "Envelope_DeleteResultData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteResultData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[DeleteResultData]"
      },
...
      "Envelope_PortfolioDTO_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PortfolioDTO"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[PortfolioDTO]"
      },
      "Envelope_PortfolioListData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PortfolioListData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[PortfolioListData]"
      },
      "Envelope_PortfolioPositionMutationData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PortfolioPositionMutationData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[PortfolioPositionMutationData]"
      },
      "Envelope_PortfolioPositionsData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PortfolioPositionsData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[PortfolioPositionsData]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          },
          "total_assets": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Assets"
          },
          "available_cash": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Cash"
          },
          "risk_per_trade_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Per Trade Pct"
          },
          "min_risk_reward_ratio": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Risk Reward Ratio"
          },
          "max_total_position_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Total Position Pct"
          },
          "max_single_stock_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Single Stock Pct"
          },
          "max_sector_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Sector Pct"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioCreateRequest"
      },
      "PortfolioDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "position_count": {
            "type": "integer",
            "title": "Position Count"
          },
          "total_assets": {
            "type": "number",
            "title": "Total Assets"
          },
          "available_cash": {
            "type": "number",
            "title": "Available Cash"
          },
          "risk_per_trade_pct": {
            "type": "number",
            "title": "Risk Per Trade Pct"
          },
          "min_risk_reward_ratio": {
            "type": "number",
            "title": "Min Risk Reward Ratio"
          },
          "max_total_position_pct": {
            "type": "number",
            "title": "Max Total Position Pct"
          },
          "max_single_stock_pct": {
            "type": "number",
            "title": "Max Single Stock Pct"
          },
          "max_sector_pct": {
            "type": "number",
            "title": "Max Sector Pct"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "position_count",
          "total_assets",
          "available_cash",
          "risk_per_trade_pct",
          "min_risk_reward_ratio",
          "max_total_position_pct",
          "max_single_stock_pct",
          "max_sector_pct",
          "created_at",
          "updated_at"
        ],
        "title": "PortfolioDTO"
      },
      "PortfolioListData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PortfolioDTO"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PortfolioListData"
      },
      "PortfolioPositionDTO": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "market": {
            "type": "string",
            "enum": [
              "US",
              "KR",
              "CN"
            ],
            "title": "Market"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "average_cost": {
            "type": "number",
            "title": "Average Cost"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "market",
          "symbol",
          "quantity",
          "average_cost",
          "updated_at"
        ],
        "title": "PortfolioPositionDTO"
      },
      "PortfolioPositionMutationData": {
        "properties": {
          "position": {
            "$ref": "#/components/schemas/PortfolioPositionDTO"
          },
          "portfolio_revision": {
            "type": "integer",
            "title": "Portfolio Revision"
          }
        },
        "type": "object",
        "required": [
          "position",
          "portfolio_revision"
        ],
        "title": "PortfolioPositionMutationData"
      },
      "PortfolioPositionsData": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "portfolio_revision": {
            "type": "integer",
            "title": "Portfolio Revision"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PortfolioPositionDTO"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "portfolio_revision",
          "items"
        ],
        "title": "PortfolioPositionsData"
      },
...
      "PortfolioUpdateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          },
          "total_assets": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Assets"
          },
          "available_cash": {
            "type": "number",
            "minimum": 0.0,
            "title": "Available Cash"
          },
          "risk_per_trade_pct": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Risk Per Trade Pct"
          },
          "min_risk_reward_ratio": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Min Risk Reward Ratio"
          },
          "max_total_position_pct": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Total Position Pct"
          },
          "max_single_stock_pct": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Single Stock Pct"
          },
          "max_sector_pct": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Sector Pct"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "total_assets",
          "available_cash",
          "risk_per_trade_pct",
          "min_risk_reward_ratio",
          "max_total_position_pct",
          "max_single_stock_pct",
          "max_sector_pct",
          "expected_version"
        ],
        "title": "PortfolioUpdateRequest",
        "description": "原子更新名称与全部账户字段（plan 4.2.1：一次条件更新、成功仅 version+1）。"
      },
      "PositionUpsertRequest": {
        "properties": {
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "average_cost": {
            "type": "number",
            "title": "Average Cost"
          },
          "expected_portfolio_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Portfolio Revision"
          }
        },
        "type": "object",
        "required": [
          "quantity",
          "average_cost",
          "expected_portfolio_revision"
        ],
        "title": "PositionUpsertRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}